+++ apps/api/src/modules/playlist/playlist.service.ts
@@ -55,9 +55,13 @@ export class PlaylistService {
   }
 
   async getPlaylistDetail(
+    userId: string,
     playlistId: string,
   ): Promise<GetPlaylistDetailResDto> {
-    const playlist = await this.playlistRepo.getPlaylistDetail(playlistId);
+    const playlist = await this.playlistRepo.getPlaylistDetail(
+      userId,
+      playlistId,
+    );
     if (!playlist)
       throw new NotFoundException('플레이리스트가 존재하지 않습니다.');
 
@@ -84,8 +88,12 @@ export class PlaylistService {
     };
   }
 
-  async update(playlistId: string, title: string): Promise<Playlist> {
-    const playlist = await this.playlistRepo.findById(playlistId);
+  async update(
+    userId: string,
+    playlistId: string,
+    title: string,
+  ): Promise<Playlist> {
+    const playlist = await this.playlistRepo.findById(userId, playlistId);
     if (!playlist)
       throw new NotFoundException('플레이리스트가 존재하지 않습니다.');
 
@@ -94,17 +102,22 @@ export class PlaylistService {
     return this.playlistRepo.save(playlist);
   }
 
-  async delete(playlistId: string): Promise<void> {
-    const isDeleted = await this.playlistRepo.delete(playlistId);
+  async delete(userId: string, playlistId: string): Promise<void> {
+    const isDeleted = await this.playlistRepo.delete(userId, playlistId);
 
     if (!isDeleted)
       throw new NotFoundException('플레이리스트가 존재하지 않습니다.');
   }
 
   async addMusics(
+    userId: string,
     playlistId: string,
     musics: MusicRequestDto[],
   ): Promise<void> {
+    const playlist = await this.playlistRepo.findById(userId, playlistId);
+    if (!playlist)
+      throw new NotFoundException('플레이리스트가 존재하지 않습니다.');
+
     const musicIds = await Promise.all(
       musics.map(async (m) => {
         if (m.id) return m.id;
@@ -121,7 +134,15 @@ export class PlaylistService {
     });
   }
 
-  async changeMusicOrder(playlistId: string, musicIds: string[]) {
+  async changeMusicOrder(
+    userId: string,
+    playlistId: string,
+    musicIds: string[],
+  ) {
+    const playlist = await this.playlistRepo.findById(userId, playlistId);
+    if (!playlist)
+      throw new NotFoundException('플레이리스트가 존재하지 않습니다.');
+
     return await this.ds.transaction(async (tx) => {
       await this.playlistRepo.deleteAllMusics(playlistId, tx);
       await this.playlistRepo.addMusics(playlistId, musicIds, 0, tx);
